Scope down manifest interface
